+++ Emerald/src/graphics/window.h
@@ -23,8 +23,8 @@ public:
 	inline void SetWidth(uint width) { m_width = width; }
 	inline void SetHeight(uint height) { m_height = height; }
 
-	inline int GetWidth() { return m_width; }
-	inline int GetHeight() { return m_height; }
+	inline float GetWidth() { return (float)m_width; }
+	inline float GetHeight() { return (float)m_height; }
 	inline float GetAspect() { return (float)(m_width) / m_height; }
 
 	~Window() {
